Handle "./update.bash" without a remote
It turns out that sometimes (from the buildbot, for example) we can't
actually do a "git pull" correctly because the g it remote isn't setup
right.  This instead just ignores the "git pull" failure, assuming
that if "git pull" fails then we're as close as we can get to the
latest version.